+++ engine/server/engine/enclave_manager/enclave_manager_test.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"github.com/blang/semver"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/docker_manager/types"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/container_status"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
 	"github.com/kurtosis-tech/kurtosis-core/launcher/api_container_launcher"
 	"github.com/kurtosis-tech/object-attributes-schema-lib/schema"
@@ -63,5 +64,12 @@ func TestGetEnclaveContainersStatusFromEnclaveStatusCompleteness(t *testing.T) {
 	for _, enclaveStatus := range enclave.EnclaveStatusValues() {
 		_, err := getEnclaveContainersStatusFromEnclaveStatus(enclaveStatus)
 		require.NoError(t, err, "No EnclaveContainersStatus provided for enclave status '%v'", enclaveStatus.String())
+	}
+}
+
+func TestGetApiContainerStatusFromContainerStatusCompleteness(t *testing.T) {
+	for _, containerStatus := range container_status.ContainerStatusValues() {
+		_, err := getApiContainerStatusFromContainerStatus(containerStatus)
+		require.NoError(t, err, "No ApiContainerStatus provided for container status '%v'", containerStatus.String())
 	}
 }