+++ engine/docs/changelog.md
@@ -1,7 +1,8 @@
 # TBD
 * Updated enclave_manager to use kurtosis_backend for enclave operations
+* Updated enclave_manager to use kurtosis_backend for api container operations in an enclave
 * Removed cleanMetadataAcquisitionTestsuites from enclave_manager
-
+* Removed docker specific code from enclave_manager
 # 1.11.5
 ### Features
 * Use container-engine-lib 0.10.1 and core 1.39.9 to provide API container & enclave CRUD functions